move regex method to Util and start to factor out text input expansion
