+++ rabbit-escape-ui-swing/src/rabbitescape/ui/swing/GitHubJsonTools.java
@@ -75,29 +75,4 @@ public class GitHubJsonTools
         return ret.toArray( values );
     }
 
-    /**
-     * @brief Replace all matches, but preserves the first capturing group.
-     *        Compare to String.replaceAll, which replaces the whole match. This
-     *        is in the JSON class because JSON String vales have escaped
-     *        characters to be cleaned up.
-     * @param patternFlags The flags from java.util.regex.Pattern.
-     */
-    public static String regexReplacePreserveGroup( String s, String regex, int patternFlags )
-    {
-        Pattern p = Pattern.compile( regex, patternFlags );
-        Matcher dsMatcher = p.matcher( s );
-        StringBuffer sb = new StringBuffer();
-        while ( dsMatcher.find() )
-        {
-            dsMatcher.appendReplacement( sb, "$1" );
-        }
-        dsMatcher.appendTail( sb );
-
-        return sb.toString();
-    }
-
-    public static String regexReplacePreserveGroup( String s, String regex )
-    {
-        return regexReplacePreserveGroup( s, regex, 0 );
-    }
 }
